+++ lib.profiler/src/org/netbeans/lib/profiler/heap/ClassDumpSegment.java
@@ -171,8 +171,8 @@ class ClassDumpSegment extends TagBounds {
     
     void addInstanceSize(ClassDump cls, int tag, long instanceOffset) {
         if ((tag == HprofHeap.OBJECT_ARRAY_DUMP) || (tag == HprofHeap.PRIMITIVE_ARRAY_DUMP)) {
-            Integer sizeInt = (Integer) arrayMap.get(cls);
-            int size = 0;
+            Long sizeInt = (Long) arrayMap.get(cls);
+            long size = 0;
             HprofByteBuffer dumpBuffer = hprofHeap.dumpBuffer;
             int idSize = dumpBuffer.getIDSize();
             long elementsOffset = instanceOffset + 1 + idSize + 4;
@@ -190,8 +190,8 @@ class ClassDumpSegment extends TagBounds {
                 elSize = idSize;
             }
 
-            size += (getMinimumInstanceSize() + (elements * elSize));
-            arrayMap.put(cls, Integer.valueOf(size));
+            size += (getMinimumInstanceSize() + (((long)elements) * elSize));
+            arrayMap.put(cls, Long.valueOf(size));
         }
     }
